7mo CICLO/SISTEMAS MULTIAGENTES/Agente con Slack/agente_IA_30min/requirements.txt
llama3-package
langchain
langchain_ollama
python-dotenv
slack_bolt